requirements.txt
django 
djangorestframework 
Pillow 
tensorflow 
scikit-learn 
numpy 
tqdm
psycopg2-binary